missing param + code rationalisation
